Optimise day 6 part 2 solution (10x speedup)
My part 2 solution already contained 2 optimisations: (1) pre-compute a "visited" path in part 1 and only try placing obstacles on nodes explored in our original path, and (2) restore simulations from just before the point of collision.

I add a third optimisation which results in around 10x speedup. For each (pos, dir) we visit, we compute and memoize the location of the next obstacle (or grid boundary) it hits. When we place an obstacle in part 2, we always progress to the less boundary, unless the obstacle was placed between our current position and the next position - at which point we instead move to just before the obstacle. This allows us to memoize long repeated iterations of walking, which happens a lot as cycles on the path tend to form in the same place. Hence cutting runtime by a further 10x or so (around 0.3 seconds on Python and milliseconds on Rust)