Optimise day 21 further
Gives about 1.1x speedup (-10% to 11%).
The best path will always be either all vertical moves then all horizontal moves, or all horizontal moves then all vertical moves. All other paths (which I was computing with combinatorics) need not be explored.

Memoisation means this doesn't result in a huge speedup but still cuts a further 10% or so of the runtime.